# docs/api/openapi-mvp3-draft.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontology Platform Backend",
    "description": "Ontology Platform MVP backend API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/quality/summary": {
      "get": {
        "tags": [
          "Quality"
        ],
        "summary": "Get quality summary v0.1",
        "operationId": "get_quality_summary_api_v1_projects__project_id__quality_summary_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "ontology_version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ontology Version Id"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "name": "extraction_job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Extraction Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualitySummary"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QualityCandidateCounts": {
        "properties": {
          "total": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "entity": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "relation": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "property_value": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "missing_evidence": {
            "$ref": "#/components/schemas/QualityCountMetric"
          }
        },
        "type": "object",
        "required": [
          "total",
          "entity",
          "relation",
          "property_value",
          "missing_evidence"
        ],
        "title": "QualityCandidateCounts"
      },
      "QualityCountMetric": {
        "properties": {
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "drilldown": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityDrilldownHint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "QualityCountMetric"
      },
      "QualityDrilldownHint": {
        "properties": {
          "target": {
            "$ref": "#/components/schemas/QualityDrilldownTarget"
          },
          "query": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "target",
          "query"
        ],
        "title": "QualityDrilldownHint"
      },
      "QualityDrilldownTarget": {
        "type": "string",
        "enum": [
          "review_inbox",
          "publish_jobs",
          "published_graph"
        ],
        "title": "QualityDrilldownTarget"
      },
      "QualityPublishCounts": {
        "properties": {
          "not_published": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "published": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "rolled_back": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "published_entities": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "published_relations": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "publish_success": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "publish_failed": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "current_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version"
          }
        },
        "type": "object",
        "required": [
          "not_published",
          "published",
          "rolled_back",
          "published_entities",
          "published_relations",
          "publish_success",
          "publish_failed"
        ],
        "title": "QualityPublishCounts"
      },
      "QualityRateMetric": {
        "properties": {
          "numerator": {
            "type": "integer",
            "title": "Numerator"
          },
          "denominator": {
            "type": "integer",
            "title": "Denominator"
          },
          "rate": {
            "type": "number",
            "title": "Rate"
          },
          "drilldown": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityDrilldownHint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "numerator",
          "denominator",
          "rate"
        ],
        "title": "QualityRateMetric"
      },
      "QualityRates": {
        "properties": {
          "approval_rate": {
            "$ref": "#/components/schemas/QualityRateMetric"
          },
          "rejection_rate": {
            "$ref": "#/components/schemas/QualityRateMetric"
          },
          "modification_rate": {
            "$ref": "#/components/schemas/QualityRateMetric"
          },
          "validation_failure_rate": {
            "$ref": "#/components/schemas/QualityRateMetric"
          },
          "evidence_missing_rate": {
            "$ref": "#/components/schemas/QualityRateMetric"
          },
          "published_ratio": {
            "$ref": "#/components/schemas/QualityRateMetric"
          }
        },
        "type": "object",
        "required": [
          "approval_rate",
          "rejection_rate",
          "modification_rate",
          "validation_failure_rate",
          "evidence_missing_rate",
          "published_ratio"
        ],
        "title": "QualityRates"
      },
      "QualityReviewCounts": {
        "properties": {
          "pending": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "approved": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "rejected": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "modified": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "needs_discussion": {
            "$ref": "#/components/schemas/QualityCountMetric"
          }
        },
        "type": "object",
        "required": [
          "pending",
          "approved",
          "rejected",
          "modified",
          "needs_discussion"
        ],
        "title": "QualityReviewCounts"
      },
      "QualitySummary": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "ontology_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ontology Version Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "candidate_counts": {
            "$ref": "#/components/schemas/QualityCandidateCounts"
          },
          "validation_counts": {
            "$ref": "#/components/schemas/QualityValidationCounts"
          },
          "review_counts": {
            "$ref": "#/components/schemas/QualityReviewCounts"
          },
          "publish_counts": {
            "$ref": "#/components/schemas/QualityPublishCounts"
          },
          "rates": {
            "$ref": "#/components/schemas/QualityRates"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "generated_at",
          "candidate_counts",
          "validation_counts",
          "review_counts",
          "publish_counts",
          "rates"
        ],
        "title": "QualitySummary"
      },
      "QualityValidationCounts": {
        "properties": {
          "not_validated": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "passed": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "warning": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "failed": {
            "$ref": "#/components/schemas/QualityCountMetric"
          },
          "by_rule_code": {
            "additionalProperties": {
              "$ref": "#/components/schemas/QualityCountMetric"
            },
            "type": "object",
            "title": "By Rule Code"
          }
        },
        "type": "object",
        "required": [
          "not_validated",
          "passed",
          "warning",
          "failed",
          "by_rule_code"
        ],
        "title": "QualityValidationCounts"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}